feat: supported essay (#42)

+++ plugins/docusaurus-terminology/src/index.ts
@@ -1,8 +1,8 @@
 import fs from "fs";
 import path from "path";
 import type { Plugin } from "@docusaurus/types";
-import type { Configuration, RuleSetRule, RuleSetUseItem } from "webpack";
-
+import type { RuleSetRule, RuleSetUseItem } from "webpack";
+import yaml from "js-yaml";
 export interface DocusaurusContext {
   baseUrl: string;
   siteDir: string;
@@ -24,64 +24,65 @@ export interface ConfigureWebpackUtils {
 }
 
 export interface TerminologyOptions {
-  baseUrl?: string;
-  resolved?: boolean;
+  path: string;
+  glossaries: string;
+  routeBasePath?: string;
   glossaryComponentPath?: string;
+}
+
+export interface Terminology {
+  slug: string;
+  name: string;
+  description: string;
+  authors: string[];
+  path: string;
 }
 
 export default async function DocusaurusTerminologyPlugin(
   context: DocusaurusContext,
   options: TerminologyOptions,
-): Promise<Plugin<any>> {
+): Promise<Plugin<{ terminologies: Record<string, Terminology> }>> {
   try {
-    fs.rm("node_modules/.cache", { recursive: true }, (err) => {
+    fs.stat("node_modules/.cache", (err, stats) => {
       if (err) {
         console.error(err);
+        return;
+      } else if (stats.isDirectory()) {
+        fs.rm("node_modules/.cache", { recursive: true }, (err) => {
+          if (err) {
+            console.error(err);
+          }
+        });
       }
     });
-  } catch (err) {}
+  } catch (err) {
+    console.error(err);
+  }
   return {
     name: "terminology-docusaurus-plugin",
     configureWebpack(config, isServer, utils, content) {
-      options.baseUrl = config.output?.publicPath as string;
-      options.resolved = true;
-
       const rules = config.module?.rules as WebpackRule[];
       let rule = rules.find((rule) => {
+        if (!options.path) {
+          return false;
+        }
+        const targetPath = path.resolve(process.cwd(), options.path);
         return (
+          rule.include &&
+          Array.isArray(rule.include) &&
+          rule.include.some((include) =>
+            include.toString().includes(targetPath),
+          ) &&
           rule.use &&
           Array.isArray(rule.use) &&
           rule.use.some(
             (kider: RuleSetUseItem) =>
               typeof kider === "object" &&
               typeof kider.loader === "string" &&
-              kider.loader.includes("plugin-content-blog"),
+              kider.loader.includes("mdx-loader"),
           )
         );
       });
-
-      if (!rule) {
-        rule = rules.find((rule) => {
-          return (
-            rule.use &&
-            Array.isArray(rule.use) &&
-            rule.use.some(
-              (kider: RuleSetUseItem) =>
-                typeof kider === "object" &&
-                typeof kider.loader === "string" &&
-                kider.loader.includes("mdx-loader"),
-            )
-          );
-        });
-      }
-
-      if (!rule) {
-        rule = rules.find((rule) => {
-          if (!rule.test) return false;
-          const ruleRegExp = new RegExp(rule.test as string);
-          return ruleRegExp.test("test.md") && ruleRegExp.test("test.mdx");
-        });
-      }
 
       if (
         rule &&
@@ -108,6 +109,35 @@ export default async function DocusaurusTerminologyPlugin(
         mergeStrategy: { module: "replace" },
         module: config.module,
       };
+    },
+    async loadContent() {
+      const terminologiesPath = path.resolve(
+        context.siteDir,
+        options.glossaries,
+      );
+      const terminologies = yaml.load(
+        fs.readFileSync(terminologiesPath, "utf8"),
+      ) as Record<string, Terminology>;
+      Object.entries(terminologies).forEach(([key, terminology]) => {
+        terminology.path = path.join(
+          options.path || "terminology",
+          key.toLowerCase().replace(/ /g, "-"),
+        );
+        terminology.slug = path.join(
+          options.routeBasePath || "terms",
+          key.toLowerCase().replace(/ /g, "-"),
+        );
+      });
+      return {
+        terminologies,
+      };
+    },
+    async contentLoaded({ content, actions }) {
+      actions.setGlobalData({
+        terminologies: (
+          content as { terminologies: Record<string, Terminology> }
+        ).terminologies,
+      });
     },
   };
 }
